requirements.txt
Django==3.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.4
clearbit==0.1.7
